refactor(tag): 收敛标签权限弹窗逻辑
